+++ edgy/conf/__init__.py
@@ -1,16 +1,14 @@
 import os
-from typing import Any, Optional, Type
+from typing import Any
 
 from edgy.conf.functional import LazyObject, empty
 from edgy.conf.module_import import import_string
 
 ENVIRONMENT_VARIABLE = "EDGY_SETTINGS_MODULE"
 
-DBSettings = Type["EdgyLazySettings"]
-
 
 class EdgyLazySettings(LazyObject):
-    def _setup(self, name: Optional[str] = None) -> None:
+    def _setup(self) -> None:
         """
         Load the settings module pointed to by the environment variable. This
         is used the first time settings are needed, if the user hasn't
@@ -26,7 +24,7 @@ class EdgyLazySettings(LazyObject):
 
         self._wrapped = settings()
 
-    def __repr__(self: "EdgyLazySettings") -> str:
+    def __repr__(self) -> str:
         # Hardcode the class name as otherwise it yields 'Settings'.
         if self._wrapped is empty:
             return "<EdgyLazySettings [Unevaluated]>"
@@ -38,4 +36,7 @@ class EdgyLazySettings(LazyObject):
         return self._wrapped is not empty
 
 
-settings: DBSettings = EdgyLazySettings()  # type: ignore
+settings: EdgyLazySettings = EdgyLazySettings()
+
+
+__all__ = ["EdgyLazySettings", "settings"]