+++ edgy/conf/global_settings.py
@@ -1,4 +1,6 @@
+import os
 from functools import cached_property
+from pathlib import Path
 from typing import Any, Dict, List, Union
 
 from pydantic_settings import BaseSettings, SettingsConfigDict
@@ -12,13 +14,14 @@ class MediaSettings(BaseSettings):
     file_upload_temp_dir: Union[str, None] = None
     file_upload_permissions: Union[str, Any] = 0o644
     file_upload_directory_permissions: Union[str, None] = None
-    media_root: str = ""
+    # don't allow overwriting the project files by default, set to media
+    media_root: Union[str, os.PathLike] = Path("media/")
     media_url: str = ""
 
     # Storage defaults
     storages: Dict[str, dict] = {
         "default": {
-            "backend": "edgy.core.files.storage.FileSystemStorage",
+            "backend": "edgy.core.files.storage.filesystem.FileSystemStorage",
         },
         # "staticfiles": {
         #     "storage": "django.contrib.staticfiles.storage.StaticFilesStorage",