+++ edgy/core/db/models/base.py
@@ -103,7 +103,7 @@ class EdgyBaseModel(ModelParser, BaseModel, BaseModelType, metaclass=BaseModelMe
         return {k: v for k, v in kwargs.items() if k in self.meta.fields}
 
     def __repr__(self) -> str:
-        return f"<{self.__class__.__name__}: {self}>"
+        return f"<{self.__class__.__name__}: {str(self)}>"
 
     def __str__(self) -> str:
         pkl = []