+++ edgy/core/files/__init__.py
@@ -1,0 +1,4 @@
+from .base import ContentFile, FieldFile, File
+from .storage import Storage, storages
+
+__all__ = ["File", "ContentFile", "FieldFile", "Storage", "storages"]