+++ edgy/core/files/storage/__init__.py
@@ -1,7 +1,6 @@
 from .base import Storage
-from .filesystem import FileSystemStorage
 from .handler import StorageHandler
 
 storages = StorageHandler()
 
-__all__ = ["Storage", "FileSystemStorage", "StorageHandler", "storages"]
+__all__ = ["Storage", "StorageHandler", "storages"]